actualizacion

+++ database/migrations/2024_02_01_011753_update_user_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::table('users',function(Blueprint $table){
             $table->string('rol')->nullable();
             $table->string('imagen')->default('https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460_1280.png');
-            $table->bigInteger('id_google')->nullable();
+            $table->decimal('id_google', 40, 0)->nullable();
         });
     }
